made it possible to inject test decks when spawning server. separated test decks into their own module. used commandline argv to spawn a server using a specific deck. figured out how to use puppeteer's waitForSelector using then/catch to try clicking only if it appears
