+++ blackjackTable.frontend.test.js
@@ -12,7 +12,7 @@ const USER_DATA_DIR_WSL = '/mnt/c/Users/Me/Downloads/temp';
 
 const initServer = async () => {
   return new Promise((resolve, reject) => {
-    childProcess = spawn('node', ['server.js']);
+    childProcess = spawn('node', ['server.js', '--deck=dealerHasBlackjackDeck']);
 
     childProcess.stdout.on('data', data => {
       console.log(JSON.stringify(data.toString('utf8')));
@@ -21,8 +21,11 @@ const initServer = async () => {
   })
 }
 
+const killServer = () => {
+  childProcess.kill();
+}
+
 beforeAll(async () => {
-  await initServer();
 
   browser = await puppeteer.launch({
     executablePath: `chrome.exe`,
@@ -44,24 +47,28 @@ beforeAll(async () => {
 
 afterAll(() => {
   browser.close();
-  childProcess.kill();
 })
 
 test('loads blackjack page', async () => {
+  await initServer();
   await pages[0].goto(APP);
   const titleText = await pages[0].$eval('title', el => el.innerHTML);
   expect(titleText).toContain('Blackjack');
+  killServer();
 });
 
 test('player can join the game with 1000 chips', async () => {
+  await initServer();
   await pages[0].goto(APP);
   dialogValue = "1000"; // string value expected by puppeteer
   await pages[0].$eval('#joinGame', el => el.click());
   const chipsInHand = await pages[0].$eval('#chipsInHand', el => el.innerHTML);
   expect(chipsInHand).toBe('Chips: 1000');
+  killServer();
 });
 
 test('player stands, game resolves', async () => {
+  await initServer();
   await pages[0].goto(APP);
   dialogValue = "100"
   await pages[0].$eval('#joinGame', el => el.click());
@@ -71,13 +78,21 @@ test('player stands, game resolves', async () => {
   dialogValue = "10"
   await pages[0].$eval('#placeBet', el => el.click());
   await pages[0].$eval('#startRound', el => el.click());
-  await pages[0].waitForSelector('#playStand');
-  await pages[0].$eval('#playStand', el => el.click());
+  await pages[0]
+    .waitForSelector('#playStand', {timeout:200})
+    .then(async () =>  {
+      await pages[0].$eval('#playStand', el => el.click());
+    })
+    .catch((e) => {
+      console.log(e)
+    });
   const messageLog = await pages[0].$eval('div.messageLog', el => el.innerHTML);
   expect(messageLog).toContain('All bets resolved!');
-});
+  killServer();
+}, 7000);
 
 test('two players join, both players stand, game resolves', async () => {
+  await initServer();
   await pages[0].goto(APP);
   await pages[1].goto(APP);
 
@@ -99,8 +114,23 @@ test('two players join, both players stand, game resolves', async () => {
   await pages[1].$eval('#placeBet', el => el.click());
 
   await pages[0].$eval('#startRound', el => el.click());
-  await pages[0].$eval('#playStand', el => el.click());
-  await pages[1].$eval('#playStand', el => el.click());
+  await pages[0]
+    .waitForSelector('#playStand', {timeout:200})
+    .then(async () =>  {
+      await pages[0].$eval('#playStand', el => el.click());
+    })
+    .catch((e) => {
+      console.log(e)
+    });
+  await pages[1]
+    .waitForSelector('#playStand', {timeout:200})
+    .then(async () =>  {
+      await pages[1].$eval('#playStand', el => el.click());
+    })
+    .catch((e) => {
+      console.log(e)
+    });
   const messageLog = await pages[0].$eval('div.messageLog', el => el.innerHTML);
   expect(messageLog).toContain('All bets resolved!');
+  killServer();
 }, 7000);
