+++ deckConfig.js
@@ -1,0 +1,6 @@
+// configured decks for game/frontend testing
+const decks = {
+  dealerHasBlackjackDeck: require('./testDeckConfigs/dealerHasBlackjackDeck.js'),
+}
+
+module.exports = decks;
