+++ package.json
@@ -19,6 +19,7 @@
 	},
 	"dependencies": {
 		"express": "^4.16.3",
+		"minimist": "^1.2.0",
 		"socket.io": "^2.1.1"
 	}
 }
